Move everything in `stacks-codec` -> `stacks-common`
