+++ stacks-codec/Cargo.toml
@@ -13,3 +13,7 @@ description = "Binary consensus serialization codec for the Stacks blockchain"
 keywords = ["stacks", "stx", "blockchain", "codec", "serialization"]
 readme = "README.md"
 edition = "2021"
+
+[dependencies]
+stacks-common = { path = "../stacks-common", default-features = false }
+
