+++ stacks-common/Cargo.toml
@@ -28,7 +28,6 @@ name = "stacks_common"
 path = "./src/libcommon.rs"
 
 [dependencies]
-stacks-codec = { path = "../stacks-codec" }
 chrono = { version = "0.4.41", default-features = false, features = ["clock"] }
 curve25519-dalek = { version = "4.1.3", default-features = false, features = ["serde"] }
 ed25519-dalek = { workspace = true }
